+++ src/main/kotlin/com/loadingbyte/cinecred/delivery/VideoContainerRenderJob.kt
@@ -39,7 +39,7 @@ import com.loadingbyte.cinecred.imaging.ColorSpace.Transfer.Companion.PQ
 import com.loadingbyte.cinecred.imaging.ColorSpace.Transfer.Companion.SRGB
 import com.loadingbyte.cinecred.imaging.DeferredImage.Companion.STATIC
 import com.loadingbyte.cinecred.imaging.DeferredImage.Companion.TAPES
-import com.loadingbyte.cinecred.project.Project
+import com.loadingbyte.cinecred.project.Styling
 import org.bytedeco.ffmpeg.global.avcodec.*
 import org.bytedeco.ffmpeg.global.avutil.*
 import java.nio.file.Path
@@ -50,7 +50,7 @@ import kotlin.math.pow
 class VideoContainerRenderJob private constructor(
     private val format: Format,
     private val config: Config,
-    private val project: Project,
+    private val styling: Styling,
     private val video: DeferredVideo,
     private val file: Path
 ) : RenderJob {
@@ -81,7 +81,7 @@ class VideoContainerRenderJob private constructor(
         val colorSpace = if (matte) ColorSpace.of(BT709, LINEAR) else ColorSpace.of(config[PRIMARIES], config[TRANSFER])
         val ceiling = if (colorSpace.transfer.isHDR) null else 1f
         val scan = config[SCAN]
-        val grounding = if (config[TRANSPARENCY] == GROUNDED) project.styling.global.grounding else null
+        val grounding = if (config[TRANSPARENCY] == GROUNDED) styling.global.grounding else null
         val scaledVideo = video.copy(2.0.pow(config[RESOLUTION_SCALING_LOG2]), config[FPS_SCALING])
 
         val writerSpec = Bitmap.Spec(
@@ -229,12 +229,12 @@ class VideoContainerRenderJob private constructor(
 
         override fun createRenderJob(
             config: Config,
-            project: Project,
+            styling: Styling,
             pageDefImages: List<DeferredImage>?,
             video: DeferredVideo?,
             fileOrDir: Path,
             filenamePattern: String?
-        ) = VideoContainerRenderJob(this, config, project, video!!, fileOrDir)
+        ) = VideoContainerRenderJob(this, config, styling, video!!, fileOrDir)
 
     }